Add ProductSyncer and other related services
Several new services/classes are introduced to handle the process of syncing a product with Google's Merchant Center.

The ProductSyncer service is the main hub for other classes to use when they need to update or delete a product to or from GMC.

The GoogleProductService is a wrapper for the Google's Content API client to simplify the product related business logic for our use case.

The update/delete is be done in batches of 1000 per request (recommended by Google). There is no queue/async service implemented yet. The synchronous version of this functionality can be tested in the ConnectionTest page.

When a product is uploaded to Google, 2 new metadata values are created for it (`synced_at`, `google_id`) which indicate the product's sync status. If the product is deleted/disabled/hidden, those values will be deleted too, meaning the product is not synced to GMC.