Fix UPDATE_SHIPPING_RATE for when there are no existing rates.
When there are no existing rates, the modal is prepopulated with all countries by default. Upon saving, the code would treat it as if those values exist and will call UPDATE operation, but there are no values in existing array.

The fix here is to check if there are no existing values, we would just add them into the array.