Add ActionScheduler jobs for product sync
Adds several ActionScheduler jobs to handle updating and deleting products to/from Google Merchant Center.

The jobs are scheduled automatically when products are created, updated or deleted.

There are also two jobs that handle updating or deleting ALL of the shop's products. These must be manually triggered to run (and their triggers have not been implemented in this commit).

The main plugin is also now configured to be registered on the `plugins_loaded` hook instead of the `woocommerce_loaded` hook because using the `woocommerce_loaded` introduced some problems with dependency injection (specifically loading ActionScheduler related classes).