Add selective product sync feature
User can now choose whether to sync a product with Google merchant center or not using a meta box on the product edit page.

The sync status of the product is also displayed in that box. It has placeholders for loading the list of product content issues and suggestions (sent by Google) which is yet to be implemented.