Use an incremental index as batchID
This will allow each WooCommerce product to be submitted multiple times in the same batch, which will help in cases where the same product needs to be submitted for multiple languages/countries. Or in cases where the same WooCommerce product has multiple Google assigned IDs and they all need to be deleted in the same batch.