Change back to using manual apiFetch to get billing-status.
The previous approach is using useGoogleAdsAccountBillingStatus hook. However, the parent component SetupBilling is also using the same hook to conditionally render SetupCard. This means the onStatusApproved callback in useAutoCheckBillingStatusEffect will not be called because SetupCard would have already been unmounted.

With this new approach, we are ensuring that the apiFetch, onStatusApproved and receiveGoogleAdsAccountBillingStatus are all called within one "atomic" function, instead of relying on effects which depends on the next cycle.