Compare the product's ID instead of object for more accurate results
The product objects might be modified by WooCommerce halfway through execution and in the ProductRepositoryTest it's safer to compare their IDs instead of the objects. Test results would be the same because we only expect the methods to return the product we requested (based on our query) and we don't care about the object that is returned or what it contains (that's up to WooCommerce internals).