Add root files and directories to version replace path
This change is needed for woorelease to be able to pick up these files and replace the `x.x.x` version string in them to the actual version string used for the release.

So far we only included the `./src` folder, which meant that if we had a new function or class in any other directory, woorelease wouldn't had replaced their `@since` tags with the right version.