Fetch the product from DB when hooking to create product actions
Instead of relying on the product object that WooCommerce provides to the callbacks of `woocommerce_new_product` and `woocommerce_new_product_variation` functions, make a call to the database and fetch the product using its ID.

Some product attributes seems to be only set it in the `WP_Product` object once WooCommerce persists the product into database or when the `WP_Product` is created by fetching the product data from database.

For example, the `catalog_visibility` attribute was not set for the `WC_Product` object returned by the `woocommerce_new_product_variation` action.

This was causing a notice to appear in the debug logs whenever a new product variation was created:

```
PHP Notice:  Undefined index: catalog_visibility in ./woocommerce/includes/class-wc-product-variation.php on line 419
```