Add method to monitor job timeout and retry.
This method uses PHP's `register_shutdown_function` to register a callback to run on PHP script failure due to timeout error. There wasn't any flags set in the error to specify it's a timeout so a `strpos` search is done on the PHP fatal errors to check if the timeout error has been thrown.