Add a component state canCESPromptOpen in ProductFeed
Reason to change:

There was a bug

- Success modal is displayed, CES prompt is not shown
- Open a new tab to the product feed page, CES prompt is shown
- Go back to the original tab and dismiss the success modal, the CES
	prmopt is shown, which is what we don't want.

To solve it, moving `removeCESPromptFlagFromLocal` from
`onNoticeDismissed` and `onModalShown` to only `onNoticeShown` is not
enough, as it leads to another problem:

- Success modal is displayed, CES prompt is not shown
- Open a new tab to the product feed page, CES prompt is shown
- Click `Give feedback` but the CES modal is not displayed

This is because we change to remove the CES flag from local storage when
the CES prompt is shown. When it happens the ProductFeed will re-render
and the CES prmopt will not be rendered because the flag was removed.
Then clicking `Give feedback` won't render the modal.

To solve this we add a component state `canCESPromptOpen` in ProductFeed
component, rather than check the local storage flag directly. Then we
add the logic to update the `canCESPromptOpen` state only when its value
is false.