Send the body as JSON encoded data
This is needed because in our controllers we extract the data from the
body using get_json_params.
The function get_body_params is for retrieving form data passed in POST.
It's not a problem for most controllers because they use get_param which
checks both body and json parameters.