Cast empty arrays to objects if property type is object.
This is to avoid including an empty `array` in the API response in case the property type is `object` but it has no properties (i.e. no array key-values).

An example of this is in the `GET mc/shipping/rates` API where the `options` property of each shipping rate could be returned as an empty array.