Use a single ShippingRate class for all methods
This merged the two classes previously used to define shipping rates (ShippingRateFlat and ShippingRateFree) together. To indicate a free shipping method, we now only set its rate to 0. Moreover, `free_shipping_threshold` option is now renamed to `mini_order_amount` and applies to all shipping rates. This is to be in line with how Google represents the shipping rates.