Onboarding: Don't skip to step 2 if Ads is not connected.
This fixes an issue where a the setup flow could load in step 2 if the Ads account is switched after initially setting up MC by checking that both Ads and MC are connected prior to moving to step 2.

This adds a new method to the AdsService to check if the Ads account is connected. It does so by making sure that there is an Ads ID and that the last incomplete step is either empty, or pending 'billing', which can optionally be skipped.