Improve loading of EC success modal
This adds another side effect to get the campaigns data needed to show the correct 2nd page and removes several instances of loading UI when a selector is resolving. This also adds a new side effect to the CTA when accepting terms that will turn off polling once we've determined that the user has accepted the terms, so the Enable button will be shown.