Avoid scheduling the job to synchronize all products when initializing for the first time.
When the plugin is updated, `UpdateAllProducts` job is scheduled via `Installer` and `PluginUpdate` classes. This was originally intended for data migration during plugin upgrades. However, in version 3.2.0, it was unexpectedly scheduled for new installations as well, further causing subsequent product synchronization to be continuously blocked by the error rate threshold mechanism in `ActionSchedulerJobMonitor` class.