Reduce the cache time for product statuses and issues from 12 hours to 30 minutes.
The product statuses displayed on the plugin are synchronized from Google Merchant Center.

After product data is synchronized via the Push mode, the status cache is cleared. Because it takes some time for product statuses to be updated on Google Merchant Center, there doesn't seem to be an optimal time to synchronize the status data, and newly fetched statuses are not necessarily the most current.

Previously, the cache time was one hour, so even if the status data was outdated, it wouldn't remain so for very long. However, this was later changed to a 12-hour cache in #2329, and the cache is not cleared when data is synchronized via the API Pull mode. This results in product statuses being outdated for an extended period.

While shortening the cache time to half an hour won't make the product statuses real-time, it will at least prevent the display of outdated data for a long time.